+++ docker/entrypoint.sh
@@ -3,4 +3,4 @@
 COLOR_SUCCESS='\033[0;32m'
 NC='\033[0m'
 
-exec apache2-foreground
+