+++ src/Notification/Model/LogVendorsManagerInterface.php
@@ -19,20 +19,20 @@ interface LogVendorsManagerInterface
         $statusWalletAccount,
         $status,
         $message,
-        $nbDoc
+        $nbDoc,
+        $country,
+        $paymentBlocked
     );
 
     /**
      * Insert more data into the vendor.
      * Do not save it.
      *
-     * @param VendorInterface $vendor
-     * @param array           $miraklData
+     * @param LogVendorsInterface $logVendors
+     * @param array $logData
+     * @return mixed
      */
-    public function update(
-        LogVendorsInterface $logVendors,
-        array $logData
-    );
+    public function update(LogVendorsInterface $logVendors, array $logData);
 
     /**
      * Save an array of vendors.